requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=1.0.0
langchain-openai>=0.2.0       # only needed if using OpenAI
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0